Adjusted scene to remove sound management duplication. Removed old sounds and added new ones.

+++ Assets/Scripts/Playmode/Util/Values/CRandom.cs
@@ -4,7 +4,7 @@
 	{
 		private static readonly System.Random CRand = new System.Random();
 
-		public static float Next(int minValue, int maxValue)
+		public static int Next(int minValue, int maxValue)
 		{
 			return CRand.Next(minValue, maxValue);
 		}
